Fix clang warnings.

+++ addons/TMVAHelper/src/TMVAHelper.cc
@@ -22,7 +22,7 @@ ROOT::VecOps::RVec<float>
 tmva_helper_xgb::operator()(const ROOT::VecOps::RVec<float> vars) {
   auto const tbb_slot =
       std::max(tbb::this_task_arena::current_thread_index(), 0);
-  if (tbb_slot >= m_interpreters.size()) {
+  if (tbb_slot >= static_cast<int>(m_interpreters.size())) {
     throw std::runtime_error(
         "Not enough interpreters allocated for number of tbb threads");
   }
