+++ analyzers/dataframe/src/WeaverUtils.cc
@@ -46,7 +46,7 @@ namespace FCCAnalyses {
         throw std::runtime_error("Invalid index requested for object weight " + std::to_string(weight) + ".");
       ROOT::VecOps::RVec<float> out;
       for (const auto& object_weights : objects_weights) {
-        if (weight >= object_weights.size())
+        if (weight >= static_cast<int>(object_weights.size()))
           throw std::runtime_error("Flavour weight index exceeds the number of weights registered.");
         out.emplace_back(object_weights.at(weight));
       }
